unified logic for cmdstan install dir, add fn validate_dirs to utils, reorg utils code - classes for temp files at end
